refactor(config): decompose provider.rs + engine.rs into submodules (#1200)
TD-015 + TD-016. Behaviour-preserving decomposition of the two oversized
config files into directories. Public API identical: lib.rs's
`pub use provider::*` / `pub use engine::*` re-exports flatten the same
surface because each mod.rs re-exports all its submodules' public items.

- provider.rs (2,468 LOC) -> provider/{mod,category,schema,registry,resolution}.rs
  category: InferenceCategory; schema: ProviderKind/PresetBaseUrls/ProviderPreset/
  ProviderMod/Provider/unified_memory_bytes; registry: ProviderRegistry +
  registry()/ensure_mods_loaded(); resolution: ProviderConfig/CloudConfig/
  Cli*Overrides/resolve_config/resolve_cloud_config. Tests stay in mod.rs.
- engine.rs (1,655 LOC) -> engine/{mod,session,inference,encounters,npc,reactions,
  palette,world,map}.rs split by [engine.*] config domain. mod.rs owns the
  aggregate EngineConfig + loaders + tests.

145 config tests pass; clippy -D warnings clean; arch_fitness green.
