--- parish/crates/parish-config/src/engine.rs
@@ -1,1655 +1,0 @@
-//! Engine configuration structs for tunable parameters.
-//!
-//! Every struct derives `Deserialize` and has a `Default` implementation
-//! that returns the original hardcoded values, ensuring backward compatibility
-//! when no config file is present.
-//!
-//! These are ENGINE-LEVEL parameters (timeouts, game mechanics, palette tuning).
-//! Game-specific CONTENT (prompts, loading phrases, encounter text) lives in
-//! the mod system (`GameMod` / `mod.toml`).
-
-use crate::provider::InferenceCategory;
-use parish_types::SpeedConfig;
-use serde::{Deserialize, Serialize};
-use std::collections::BTreeMap;
-use std::path::{Path, PathBuf};
-
-/// Resolves the path to `parish.toml` by walking up from `start` looking for
-/// an existing `parish.toml` file (up to 5 ancestor directories). If none is
-/// found, returns `start.join("parish.toml")` so the caller still gets a path
-/// (which will produce defaults when passed to [`load_engine_config`]).
-///
-/// Must be called once at startup with a deliberately resolved starting
-/// directory — never from a request handler or per-call helper (see Rule 9).
-pub fn resolve_config_path(start: &Path) -> PathBuf {
-    let mut p = Some(start.to_path_buf());
-    for _ in 0..5 {
-        if let Some(ref dir) = p {
-            let candidate = dir.join("parish.toml");
-            if candidate.is_file() {
-                return candidate;
-            }
-            p = dir.parent().map(|d| d.to_path_buf());
-        }
-    }
-    start.join("parish.toml")
-}
-
-/// Loads the `[engine]` section from a `parish.toml` at the given path.
-///
-/// Returns [`EngineConfig::default`] if the file is missing, unreadable, or
-/// doesn't contain a parseable `[engine]` table. After loading, calls
-/// [`MapConfig::apply_defaults`] so partial `[engine.map.tile_sources.*]`
-/// overrides don't wipe the baked-in registry.
-///
-/// Intended for Tauri/web-server boot; the CLI already has its own
-/// `resolve_config` pipeline for provider/cloud config.
-pub fn load_engine_config(path: &Path) -> EngineConfig {
-    #[derive(Deserialize, Default)]
-    struct Wrapper {
-        #[serde(default)]
-        engine: EngineConfig,
-    }
-
-    let text = match std::fs::read_to_string(path) {
-        Ok(s) => s,
-        Err(_) => return EngineConfig::default(),
-    };
-    let mut engine = toml::from_str::<Wrapper>(&text)
-        .map(|w| w.engine)
-        .unwrap_or_default();
-    engine.map.apply_defaults();
-    engine
-}
-
-/// Root engine configuration parsed from `[engine]` section of `parish.toml`.
-#[derive(Debug, Default, Deserialize, Clone)]
-pub struct EngineConfig {
-    /// LLM inference timeouts.
-    #[serde(default)]
-    pub inference: InferenceConfig,
-    /// Game speed presets.
-    #[serde(default)]
-    pub speeds: SpeedConfig,
-    /// Encounter probability by time of day.
-    #[serde(default)]
-    pub encounters: EncounterConfig,
-    /// NPC memory, cognition, and relationship tuning.
-    #[serde(default)]
-    pub npc: NpcConfig,
-    /// Color palette contrast configuration.
-    #[serde(default)]
-    pub palette: PaletteConfig,
-    /// World graph tuning.
-    #[serde(default)]
-    pub world: WorldConfig,
-    /// Persistence / save system tuning.
-    #[serde(default)]
-    pub persistence: PersistenceConfig,
-    /// Session and pacing timeouts.
-    #[serde(default)]
-    pub session: SessionConfig,
-    /// Map tile source registry and active default.
-    #[serde(default)]
-    pub map: MapConfig,
-}
-
-// ---------------------------------------------------------------------------
-// Session
-// ---------------------------------------------------------------------------
-
-/// Session and pacing timeouts.
-#[derive(Debug, Deserialize, Clone)]
-pub struct SessionConfig {
-    /// Real-time silence threshold before nearby NPCs may start banter.
-    #[serde(default = "default_idle_banter_after_secs")]
-    pub idle_banter_after_secs: u64,
-    /// Real-time inactivity threshold before the game auto-pauses.
-    #[serde(default = "default_auto_pause_after_secs")]
-    pub auto_pause_after_secs: u64,
-    /// Maximum number of concurrent in-memory sessions per server process.
-    ///
-    /// When the live session count reaches this ceiling, new session
-    /// creation is refused with `503 Service Unavailable`.  Each session
-    /// holds ~50 MB of game state (world graph, NPC manager, inference
-    /// queue), so this knob is the primary lever for the per-process memory
-    /// budget: `max_concurrent_sessions * ~50 MB ≈ memory ceiling`.
-    ///
-    /// Override at runtime with the `PARISH_MAX_SESSIONS` environment
-    /// variable (takes precedence over both the TOML setting and this
-    /// compiled-in default).
-    ///
-    /// Defaults to 50, matching the ceiling noted in issue #620.
-    #[serde(default = "default_max_concurrent_sessions")]
-    pub max_concurrent_sessions: usize,
-}
-
-impl Default for SessionConfig {
-    fn default() -> Self {
-        Self {
-            idle_banter_after_secs: default_idle_banter_after_secs(),
-            auto_pause_after_secs: default_auto_pause_after_secs(),
-            max_concurrent_sessions: default_max_concurrent_sessions(),
-        }
-    }
-}
-
-fn default_idle_banter_after_secs() -> u64 {
-    120
-}
-fn default_auto_pause_after_secs() -> u64 {
-    300
-}
-fn default_max_concurrent_sessions() -> usize {
-    50
-}
-
-// ---------------------------------------------------------------------------
-// Inference
-// ---------------------------------------------------------------------------
-
-/// LLM inference timeouts.
-#[derive(Debug, Deserialize, Clone)]
-pub struct InferenceConfig {
-    /// Non-streaming request timeout in seconds.
-    #[serde(default = "default_timeout_secs")]
-    pub timeout_secs: u64,
-    /// Streaming request timeout in seconds.
-    #[serde(default = "default_streaming_timeout_secs")]
-    pub streaming_timeout_secs: u64,
-    /// Ollama reachability check timeout in seconds.
-    #[serde(default = "default_reachability_timeout_secs")]
-    pub reachability_timeout_secs: u64,
-    /// Model download timeout in seconds.
-    #[serde(default = "default_model_download_timeout_secs")]
-    pub model_download_timeout_secs: u64,
-    /// Force Ollama setup to delete the selected local model before pulling.
-    #[serde(default)]
-    pub force_model_redownload: bool,
-    /// Model loading/warmup timeout in seconds.
-    #[serde(default = "default_model_loading_timeout_secs")]
-    pub model_loading_timeout_secs: u64,
-    /// Maximum entries in the debug inference log ring buffer.
-    #[serde(default = "default_log_capacity")]
-    pub log_capacity: usize,
-    /// Whether to also write every inference call to disk as JSONL.
-    ///
-    /// When `true` (default), each backend process writes
-    /// `{saves_dir}/inference_logs/{session}.jsonl` so users can zip the
-    /// folder and send it with a bug report. The `PARISH_INFERENCE_LOG`
-    /// env var or the `--no-inference-log` CLI flag overrides this; the
-    /// `/inference-log on|off` slash command toggles it at runtime.
-    /// API-key shapes are scrubbed before writing.
-    #[serde(default = "default_log_to_disk")]
-    pub log_to_disk: bool,
-    /// Per-category outbound request rate limits.
-    ///
-    /// Defaults to no limit. Useful when targeting paid providers
-    /// (OpenRouter, Anthropic, etc.) to avoid burning through quota
-    /// or hitting `429 Too Many Requests`.
-    #[serde(default)]
-    pub rate_limits: RateLimitConfig,
-}
-
-impl Default for InferenceConfig {
-    fn default() -> Self {
-        Self {
-            timeout_secs: default_timeout_secs(),
-            streaming_timeout_secs: default_streaming_timeout_secs(),
-            reachability_timeout_secs: default_reachability_timeout_secs(),
-            model_download_timeout_secs: default_model_download_timeout_secs(),
-            force_model_redownload: false,
-            model_loading_timeout_secs: default_model_loading_timeout_secs(),
-            log_capacity: default_log_capacity(),
-            log_to_disk: default_log_to_disk(),
-            rate_limits: RateLimitConfig::default(),
-        }
-    }
-}
-
-fn default_timeout_secs() -> u64 {
-    // Matches `default_streaming_timeout_secs` so non-streaming
-    // inference can absorb cold-loads of large local models. A 30s
-    // budget triggered timeouts on Ollama after the model unloaded
-    // post-idle (#?). Cloud APIs that respond promptly are unaffected.
-    300
-}
-fn default_streaming_timeout_secs() -> u64 {
-    300
-}
-fn default_reachability_timeout_secs() -> u64 {
-    10
-}
-fn default_model_download_timeout_secs() -> u64 {
-    3600
-}
-fn default_model_loading_timeout_secs() -> u64 {
-    300
-}
-fn default_log_capacity() -> usize {
-    50
-}
-fn default_log_to_disk() -> bool {
-    true
-}
-
-// ---------------------------------------------------------------------------
-// Rate limiting
-// ---------------------------------------------------------------------------
-
-/// Per-category rate limit configuration for outbound LLM requests.
-///
-/// All fields are optional. A `None` value disables rate limiting for
-/// that category. Categories without an explicit override fall back to
-/// the [`RateLimitConfig::default`] field, which applies to the base
-/// provider client. Configuration example (`parish.toml`):
-///
-/// ```toml
-/// [engine.inference.rate_limits.default]
-/// per_minute = 60
-/// burst = 10
-///
-/// [engine.inference.rate_limits.dialogue]
-/// per_minute = 20
-/// burst = 4
-/// ```
-#[derive(Debug, Default, Deserialize, Clone, Copy)]
-pub struct RateLimitConfig {
-    /// Default rate limit applied to the base provider client.
-    /// Categories without an explicit override share this limiter.
-    #[serde(default)]
-    pub default: Option<CategoryRateLimit>,
-    /// Override for the player-facing NPC dialogue category.
-    #[serde(default)]
-    pub dialogue: Option<CategoryRateLimit>,
-    /// Override for the background NPC simulation category.
-    #[serde(default)]
-    pub simulation: Option<CategoryRateLimit>,
-    /// Override for the player intent parsing category.
-    #[serde(default)]
-    pub intent: Option<CategoryRateLimit>,
-    /// Override for the NPC arrival reaction category.
-    #[serde(default)]
-    pub reaction: Option<CategoryRateLimit>,
-}
-
-impl RateLimitConfig {
-    /// Returns the configured rate limit for a category override, if any.
-    ///
-    /// This does NOT fall back to [`Self::default`] — the base limit is
-    /// only applied to the base client itself, not to per-category
-    /// override clients. (Override clients target a different provider
-    /// endpoint and should have their own quota.)
-    pub fn for_category(&self, cat: InferenceCategory) -> Option<CategoryRateLimit> {
-        match cat {
-            InferenceCategory::Dialogue => self.dialogue,
-            InferenceCategory::Simulation => self.simulation,
-            InferenceCategory::Intent => self.intent,
-            InferenceCategory::Reaction => self.reaction,
-        }
-    }
-}
-
-/// A single rate-limit quota: sustained rate plus burst capacity.
-///
-/// Implements a token-bucket / GCRA model: up to `burst` requests may
-/// be issued back-to-back, after which new requests are admitted at
-/// `per_minute / 60` per second until the bucket refills.
-#[derive(Debug, Deserialize, Clone, Copy)]
-pub struct CategoryRateLimit {
-    /// Sustained rate: maximum number of requests admitted per minute.
-    /// Must be greater than zero — a value of zero disables the limiter.
-    pub per_minute: u32,
-    /// Maximum burst size (token-bucket capacity). Defaults to 1.
-    #[serde(default = "default_burst")]
-    pub burst: u32,
-}
-
-fn default_burst() -> u32 {
-    1
-}
-
-// ---------------------------------------------------------------------------
-// Game Speed — SpeedConfig is defined in parish-types
-// ---------------------------------------------------------------------------
-
-// ---------------------------------------------------------------------------
-// Encounters
-// ---------------------------------------------------------------------------
-
-/// Encounter probability thresholds by time of day.
-///
-/// A random roll in `0.0..1.0` below the threshold triggers an encounter.
-#[derive(Debug, Deserialize, Clone)]
-pub struct EncounterConfig {
-    /// Encounter probability at dawn.
-    #[serde(default = "default_encounter_dawn")]
-    pub dawn: f64,
-    /// Encounter probability in the morning.
-    #[serde(default = "default_encounter_morning")]
-    pub morning: f64,
-    /// Encounter probability at midday.
-    #[serde(default = "default_encounter_midday")]
-    pub midday: f64,
-    /// Encounter probability in the afternoon.
-    #[serde(default = "default_encounter_afternoon")]
-    pub afternoon: f64,
-    /// Encounter probability at dusk.
-    #[serde(default = "default_encounter_dusk")]
-    pub dusk: f64,
-    /// Encounter probability at night.
-    #[serde(default = "default_encounter_night")]
-    pub night: f64,
-    /// Encounter probability at midnight.
-    #[serde(default = "default_encounter_midnight")]
-    pub midnight: f64,
-}
-
-impl Default for EncounterConfig {
-    fn default() -> Self {
-        Self {
-            dawn: default_encounter_dawn(),
-            morning: default_encounter_morning(),
-            midday: default_encounter_midday(),
-            afternoon: default_encounter_afternoon(),
-            dusk: default_encounter_dusk(),
-            night: default_encounter_night(),
-            midnight: default_encounter_midnight(),
-        }
-    }
-}
-
-fn default_encounter_dawn() -> f64 {
-    0.25
-}
-fn default_encounter_morning() -> f64 {
-    0.25
-}
-fn default_encounter_midday() -> f64 {
-    0.20
-}
-fn default_encounter_afternoon() -> f64 {
-    0.20
-}
-fn default_encounter_dusk() -> f64 {
-    0.15
-}
-fn default_encounter_night() -> f64 {
-    0.10
-}
-fn default_encounter_midnight() -> f64 {
-    0.05
-}
-
-// ---------------------------------------------------------------------------
-// NPC
-// ---------------------------------------------------------------------------
-
-/// NPC memory, cognition, and relationship tuning.
-#[derive(Debug, Deserialize, Clone)]
-pub struct NpcConfig {
-    /// Maximum number of entries in NPC short-term memory.
-    #[serde(default = "default_memory_capacity")]
-    pub memory_capacity: usize,
-    /// Buffer size for detecting `---` separator in streamed NPC responses.
-    #[serde(default = "default_separator_holdback")]
-    pub separator_holdback: usize,
-    /// Number of recent memories included in dialogue context.
-    #[serde(default = "default_memory_context_count")]
-    pub memory_context_count: usize,
-    /// Max characters for dialogue memory entries.
-    #[serde(default = "default_memory_truncation_dialogue")]
-    pub memory_truncation_dialogue: usize,
-    /// Max characters for event log memory entries.
-    #[serde(default = "default_memory_truncation_event_log")]
-    pub memory_truncation_event_log: usize,
-    /// Max characters for event summary in simulation.
-    #[serde(default = "default_event_summary_truncation")]
-    pub event_summary_truncation: usize,
-    /// Max characters for event summary in debug display.
-    #[serde(default = "default_event_summary_debug_truncation")]
-    pub event_summary_debug_truncation: usize,
-    /// Cognitive tier distance thresholds.
-    #[serde(default)]
-    pub cognitive_tiers: CognitiveTierConfig,
-    /// Relationship strength label thresholds.
-    #[serde(default)]
-    pub relationship_labels: RelationshipLabelConfig,
-    /// Number of recent player reactions included in dialogue context.
-    #[serde(default = "default_reaction_context_count")]
-    pub reaction_context_count: usize,
-    /// NPC arrival reaction tuning.
-    #[serde(default)]
-    pub reactions: ReactionConfig,
-    /// Hard cap on displayed NPC dialogue length in characters.
-    ///
-    /// Applied as a post-generation guard after inference completes. The
-    /// system prompt already asks for "2-4 sentences" but small models
-    /// occasionally ignore the instruction; this cap is the defensive
-    /// backstop that prevents runaway responses reaching the player (#1224).
-    /// Set to 0 to disable (not recommended in production).
-    #[serde(default = "default_dialogue_display_max_chars")]
-    pub dialogue_display_max_chars: usize,
-    /// Word-level Jaccard similarity (0.0–1.0) at or above which a fresh NPC
-    /// line is treated as a near-identical repeat of that NPC's own previous
-    /// line and replaced with a varied fallback (#1228).
-    ///
-    /// Applied as a post-generation guard after inference completes, in the
-    /// shared dialogue path. Small / quantized models occasionally echo their
-    /// previous line or loop a single clause; this deterministic backstop keeps
-    /// the degenerate output from reaching the player regardless of provider.
-    /// Intra-line clause collapse always runs; this threshold only governs the
-    /// cross-turn check. Set to 0.0 to disable the cross-turn check (not
-    /// recommended in production); 1.0 requires an identical word set.
-    #[serde(default = "default_dialogue_repetition_threshold")]
-    pub dialogue_repetition_threshold: f32,
-}
-
-impl Default for NpcConfig {
-    fn default() -> Self {
-        Self {
-            memory_capacity: default_memory_capacity(),
-            separator_holdback: default_separator_holdback(),
-            memory_context_count: default_memory_context_count(),
-            memory_truncation_dialogue: default_memory_truncation_dialogue(),
-            memory_truncation_event_log: default_memory_truncation_event_log(),
-            event_summary_truncation: default_event_summary_truncation(),
-            event_summary_debug_truncation: default_event_summary_debug_truncation(),
-            cognitive_tiers: CognitiveTierConfig::default(),
-            relationship_labels: RelationshipLabelConfig::default(),
-            reaction_context_count: default_reaction_context_count(),
-            reactions: ReactionConfig::default(),
-            dialogue_display_max_chars: default_dialogue_display_max_chars(),
-            dialogue_repetition_threshold: default_dialogue_repetition_threshold(),
-        }
-    }
-}
-
-fn default_reaction_context_count() -> usize {
-    5
-}
-
-fn default_memory_capacity() -> usize {
-    20
-}
-fn default_separator_holdback() -> usize {
-    24
-}
-fn default_memory_context_count() -> usize {
-    5
-}
-fn default_memory_truncation_dialogue() -> usize {
-    // TODO #7: 250 trimmed in-spec replies mid-sentence in the
-    // recent-events buffer fed to subsequent NPC turns. After the
-    // round-22 repetition_penalty fix, Tier 1 replies stabilise
-    // around 400–600 chars; 600 keeps them intact so downstream
-    // NPCs read the full prior dialogue, and only degenerate-loop
-    // outliers clip.
-    600
-}
-fn default_memory_truncation_event_log() -> usize {
-    // TODO #7: 150 was too tight given the dialogue cap raise.
-    // 300 leaves headroom for the outer
-    // `"{speaker} said: '{input}'. Responded: {dialogue}"` wrapper
-    // around the now-larger dialogue body without double-clipping.
-    300
-}
-fn default_event_summary_truncation() -> usize {
-    100
-}
-fn default_event_summary_debug_truncation() -> usize {
-    50
-}
-fn default_dialogue_display_max_chars() -> usize {
-    // 800 chars is roughly 5-6 sentences at typical spoken cadence.
-    // The system prompt requests "2-4 sentences"; this cap is the
-    // post-generation backstop that clips runaway model output before
-    // it reaches the player UI (#1224). Replies under ~600 chars
-    // pass through unchanged in normal operation.
-    800
-}
-fn default_dialogue_repetition_threshold() -> f32 {
-    // 0.92 word-level Jaccard: two lines must share ~92% of their word set to
-    // count as a near-identical repeat. Exact normalized equality always
-    // triggers regardless of this value. 0.92 catches the #1228 case (an NPC
-    // echoing its prior line near-verbatim) while leaving normal turn-to-turn
-    // variation — which reuses common function words but introduces new content
-    // words — comfortably below the bar. Enabled by default.
-    0.92
-}
-
-/// Cognitive tier assignment based on distance from player.
-#[derive(Debug, Deserialize, Clone)]
-pub struct CognitiveTierConfig {
-    /// Maximum distance for Tier 1 (same location).
-    #[serde(default = "default_tier1_max_distance")]
-    pub tier1_max_distance: u32,
-    /// Maximum distance for Tier 2 (nearby).
-    #[serde(default = "default_tier2_max_distance")]
-    pub tier2_max_distance: u32,
-    /// Maximum distance for Tier 3 (distant but still LLM-simulated).
-    #[serde(default = "default_tier3_max_distance")]
-    pub tier3_max_distance: u32,
-    /// Tier 2 simulation tick interval in game-minutes.
-    #[serde(default = "default_tier2_tick_interval_minutes")]
-    pub tier2_tick_interval_minutes: i64,
-    /// Tier 3 simulation tick interval in game-hours (1 game-day = 24).
-    #[serde(default = "default_tier3_tick_interval_hours")]
-    pub tier3_tick_interval_hours: i64,
-    /// Maximum NPCs per Tier 3 batch LLM call.
-    #[serde(default = "default_tier3_batch_size")]
-    pub tier3_batch_size: usize,
-    /// Tier 4 rules-engine tick interval in game-days (1 season ≈ 90 days).
-    #[serde(default = "default_tier4_tick_interval_days")]
-    pub tier4_tick_interval_days: i64,
-}
-
-impl Default for CognitiveTierConfig {
-    fn default() -> Self {
-        Self {
-            tier1_max_distance: default_tier1_max_distance(),
-            tier2_max_distance: default_tier2_max_distance(),
-            tier3_max_distance: default_tier3_max_distance(),
-            tier2_tick_interval_minutes: default_tier2_tick_interval_minutes(),
-            tier3_tick_interval_hours: default_tier3_tick_interval_hours(),
-            tier3_batch_size: default_tier3_batch_size(),
-            tier4_tick_interval_days: default_tier4_tick_interval_days(),
-        }
-    }
-}
-
-fn default_tier1_max_distance() -> u32 {
-    0
-}
-fn default_tier2_max_distance() -> u32 {
-    2
-}
-fn default_tier2_tick_interval_minutes() -> i64 {
-    5
-}
-fn default_tier3_max_distance() -> u32 {
-    5
-}
-fn default_tier3_tick_interval_hours() -> i64 {
-    24
-}
-fn default_tier3_batch_size() -> usize {
-    10
-}
-fn default_tier4_tick_interval_days() -> i64 {
-    90
-}
-
-/// Relationship strength thresholds for descriptive labels.
-#[derive(Debug, Deserialize, Clone)]
-pub struct RelationshipLabelConfig {
-    /// Threshold for "very close".
-    #[serde(default = "default_very_close")]
-    pub very_close: f64,
-    /// Threshold for "friendly".
-    #[serde(default = "default_friendly")]
-    pub friendly: f64,
-    /// Threshold for "acquainted".
-    #[serde(default = "default_acquainted")]
-    pub acquainted: f64,
-    /// Threshold for "cool".
-    #[serde(default = "default_cool")]
-    pub cool: f64,
-    /// Threshold for "strained".
-    #[serde(default = "default_strained")]
-    pub strained: f64,
-}
-
-impl Default for RelationshipLabelConfig {
-    fn default() -> Self {
-        Self {
-            very_close: default_very_close(),
-            friendly: default_friendly(),
-            acquainted: default_acquainted(),
-            cool: default_cool(),
-            strained: default_strained(),
-        }
-    }
-}
-
-fn default_very_close() -> f64 {
-    0.7
-}
-fn default_friendly() -> f64 {
-    0.3
-}
-fn default_acquainted() -> f64 {
-    0.0
-}
-fn default_cool() -> f64 {
-    -0.3
-}
-fn default_strained() -> f64 {
-    -0.7
-}
-
-// ---------------------------------------------------------------------------
-// Reactions
-// ---------------------------------------------------------------------------
-
-/// Tuning for NPC arrival reactions (greetings, nods, introductions).
-#[derive(Debug, Deserialize, Clone)]
-pub struct ReactionConfig {
-    /// Base probability that an NPC reacts when the player arrives.
-    #[serde(default = "default_reaction_base_chance")]
-    pub base_chance: f64,
-    /// Bonus when NPC is at their workplace.
-    #[serde(default = "default_reaction_workplace_bonus")]
-    pub workplace_bonus: f64,
-    /// Bonus when location is indoors.
-    #[serde(default = "default_reaction_indoor_bonus")]
-    pub indoor_bonus: f64,
-    /// Bonus when NPC has high emotional intelligence (≥4).
-    #[serde(default = "default_reaction_empathy_bonus")]
-    pub empathy_bonus: f64,
-    /// Penalty when NPC has a negative mood.
-    #[serde(default = "default_reaction_negative_mood_penalty")]
-    pub negative_mood_penalty: f64,
-    /// Penalty at night or midnight.
-    #[serde(default = "default_reaction_night_penalty")]
-    pub night_penalty: f64,
-    /// LLM timeout for reaction greeting calls (seconds).
-    #[serde(default = "default_reaction_llm_timeout_secs")]
-    pub llm_timeout_secs: u64,
-    /// Maximum number of NPCs that react on a single arrival (0 = no cap).
-    #[serde(default = "default_reaction_max_reactions")]
-    pub max_reactions: usize,
-}
-
-impl Default for ReactionConfig {
-    fn default() -> Self {
-        Self {
-            base_chance: default_reaction_base_chance(),
-            workplace_bonus: default_reaction_workplace_bonus(),
-            indoor_bonus: default_reaction_indoor_bonus(),
-            empathy_bonus: default_reaction_empathy_bonus(),
-            negative_mood_penalty: default_reaction_negative_mood_penalty(),
-            night_penalty: default_reaction_night_penalty(),
-            llm_timeout_secs: default_reaction_llm_timeout_secs(),
-            max_reactions: default_reaction_max_reactions(),
-        }
-    }
-}
-
-fn default_reaction_base_chance() -> f64 {
-    0.55
-}
-fn default_reaction_workplace_bonus() -> f64 {
-    0.35
-}
-fn default_reaction_indoor_bonus() -> f64 {
-    0.10
-}
-fn default_reaction_empathy_bonus() -> f64 {
-    0.05
-}
-fn default_reaction_negative_mood_penalty() -> f64 {
-    0.20
-}
-fn default_reaction_night_penalty() -> f64 {
-    0.15
-}
-fn default_reaction_llm_timeout_secs() -> u64 {
-    5
-}
-fn default_reaction_max_reactions() -> usize {
-    2
-}
-
-// ---------------------------------------------------------------------------
-// Palette
-// ---------------------------------------------------------------------------
-
-/// Color palette contrast configuration.
-#[derive(Debug, Deserialize, Clone)]
-pub struct PaletteConfig {
-    /// Minimum luminance contrast between foreground and background.
-    #[serde(default = "default_min_fg_bg_contrast")]
-    pub min_fg_bg_contrast: f32,
-    /// Minimum luminance contrast between muted text and background.
-    #[serde(default = "default_min_muted_bg_contrast")]
-    pub min_muted_bg_contrast: f32,
-}
-
-impl Default for PaletteConfig {
-    fn default() -> Self {
-        Self {
-            min_fg_bg_contrast: default_min_fg_bg_contrast(),
-            min_muted_bg_contrast: default_min_muted_bg_contrast(),
-        }
-    }
-}
-
-fn default_min_fg_bg_contrast() -> f32 {
-    80.0
-}
-fn default_min_muted_bg_contrast() -> f32 {
-    45.0
-}
-
-// ---------------------------------------------------------------------------
-// World
-// ---------------------------------------------------------------------------
-
-/// World graph tuning parameters.
-#[derive(Debug, Deserialize, Clone)]
-pub struct WorldConfig {
-    /// Minimum Jaro-Winkler similarity (0.0–1.0) for fuzzy location name matching.
-    ///
-    /// Higher values reduce false positives but miss more typos.
-    #[serde(default = "default_fuzzy_threshold")]
-    pub fuzzy_threshold: f64,
-}
-
-impl Default for WorldConfig {
-    fn default() -> Self {
-        Self {
-            fuzzy_threshold: default_fuzzy_threshold(),
-        }
-    }
-}
-
-fn default_fuzzy_threshold() -> f64 {
-    0.82
-}
-
-// ---------------------------------------------------------------------------
-// Persistence
-// ---------------------------------------------------------------------------
-
-/// Persistence / save system tuning parameters.
-///
-/// Currently empty — reserved for future save-system knobs (e.g. compaction,
-/// autosnap interval). The `[engine.persistence]` table is accepted for
-/// backward compatibility but has no effect.
-#[derive(Debug, Default, Deserialize, Clone)]
-pub struct PersistenceConfig {}
-
-// ---------------------------------------------------------------------------
-// Map
-// ---------------------------------------------------------------------------
-
-/// Map tile source registry and active default.
-///
-/// A registry of named raster-tile sources (XYZ templates) that the frontend
-/// can switch between at runtime via the `/map` slash command. Users can
-/// override the baked-in defaults by adding `[engine.map.tile_sources.<id>]`
-/// blocks in `parish.toml`.
-///
-/// **Partial overrides:** serde's BTreeMap deserialisation replaces the whole
-/// map rather than merging. Call [`MapConfig::apply_defaults`] after parsing
-/// to fold the baked defaults (OSM, Ireland Historic 6") back into a user-supplied
-/// registry that only overrode a subset.
-#[derive(Debug, Deserialize, Serialize, Clone)]
-pub struct MapConfig {
-    /// Id of the source used on first boot (pre-localStorage). Must match one
-    /// of the keys in `tile_sources`.
-    #[serde(default = "default_tile_source_id")]
-    pub default_tile_source: String,
-    /// Registry of available raster tile sources, keyed by id.
-    #[serde(default = "default_tile_sources")]
-    pub tile_sources: BTreeMap<String, TileSourceConfig>,
-    /// Optional path to a pre-seeded tile directory. When set, `TileCache`
-    /// checks this directory before hitting the upstream network, enabling
-    /// offline play. Path is resolved at startup; relative paths are resolved
-    /// against the mod/data directory (see CLAUDE.md rule #9). `None` means
-    /// no bundled tiles — the cache falls through to the upstream fetch as
-    /// normal. Can also be overridden at startup via `PARISH_BUNDLED_TILES_DIR`.
-    #[serde(default)]
-    pub bundled_tiles_dir: Option<std::path::PathBuf>,
-}
-
-impl Default for MapConfig {
-    fn default() -> Self {
-        Self {
-            default_tile_source: default_tile_source_id(),
-            tile_sources: default_tile_sources(),
-            bundled_tiles_dir: None,
-        }
-    }
-}
-
-impl MapConfig {
-    /// Fold baked-in defaults into the registry for any id the user didn't
-    /// override. Call this after deserialising `parish.toml` so a partial
-    /// `[engine.map.tile_sources.osm]` block doesn't wipe the historic entry.
-    pub fn apply_defaults(&mut self) {
-        for (id, source) in default_tile_sources() {
-            self.tile_sources.entry(id).or_insert(source);
-        }
-    }
-}
-
-fn default_tile_source_id() -> String {
-    "historic".to_string()
-}
-
-fn default_tile_sources() -> BTreeMap<String, TileSourceConfig> {
-    let mut m = BTreeMap::new();
-    m.insert(
-        "osm".to_string(),
-        TileSourceConfig {
-            label: "OpenStreetMap".to_string(),
-            url: "https://tile.openstreetmap.org/{z}/{x}/{y}.png".to_string(),
-            // OSM is fetched directly by the browser; no server-side proxying.
-            upstream_url: String::new(),
-            tile_size: 256,
-            minzoom: 0,
-            maxzoom: 19,
-            attribution: "© OpenStreetMap contributors".to_string(),
-            raster_saturation: -0.4,
-            raster_opacity: 0.85,
-            tms: false,
-        },
-    );
-    m.insert(
-        "historic".to_string(),
-        TileSourceConfig {
-            label: "Historic 6\" OS Ireland (1st ed., via NLS)".to_string(),
-            // Ordnance Survey of Ireland First Edition 6-inch (surveyed 1829–1842),
-            // scanned and hosted by the National Library of Scotland. NLS serves
-            // Ireland as 32 separate per-county tilesets under `os/<county>1/`
-            // — there is no seamless all-Ireland layer. Rundale is set in
-            // County Roscommon, so we wire the Roscommon 1st-edition sheet;
-            // expanding to whole-island coverage will require a multi-source
-            // style (see issue #360).
-            //
-            // Terms: CC-BY per https://maps.nls.uk/copyright.html
-            // (no version specified by NLS; CC-BY ≥ 3.0 implied). Per-sheet
-            // viewers link the licence at the #noncommercial anchor.
-            // Required attribution: "Reproduced with the permission of the
-            // National Library of Scotland". Downstream may be relicensed
-            // under CC-BY-SA per Creative Commons one-way compatibility.
-            //
-            // `url` is the same-origin proxy path the browser hits (issue #360);
-            // `upstream_url` is the absolute NLS S3 URL the server-side
-            // tile cache fetches from on a miss. The path segment after
-            // `/tiles/` must match the registering key so `tile_routes`'s
-            // validator accepts the request (PR #955).
-            url: "/tiles/historic/{z}/{x}/{y}.png".to_string(),
-            upstream_url:
-                "https://mapseries-tilesets.s3.amazonaws.com/os/roscommon1/{z}/{x}/{y}.png"
-                    .to_string(),
-            tile_size: 256,
-            minzoom: 1,
-            maxzoom: 17,
-            attribution:
-                "Historic 6\" OS Ireland (1829–1842) — Reproduced with the permission of the National Library of Scotland (CC-BY)"
-                    .to_string(),
-            raster_saturation: 0.0,
-            raster_opacity: 1.0,
-            tms: false,
-        },
-    );
-    m
-}
-
-/// A single raster tile source — URL template plus display metadata.
-#[derive(Debug, Deserialize, Serialize, Clone)]
-pub struct TileSourceConfig {
-    /// Human-readable label displayed in `/map` listings.
-    #[serde(default)]
-    pub label: String,
-    /// XYZ URL template the **frontend** uses to fetch tiles (e.g.
-    /// `https://…/{z}/{x}/{y}.png`, or a same-origin proxy path like
-    /// `/tiles/historic/{z}/{x}/{y}.png`). Empty string means the source is
-    /// registered but not yet configured; the frontend falls back to a flat
-    /// background.
-    #[serde(default)]
-    pub url: String,
-    /// Upstream XYZ URL template the **server-side tile cache** fetches from
-    /// on a cache miss (must be absolute, e.g.
-    /// `https://mapseries-tilesets.s3.amazonaws.com/os/roscommon1/{z}/{x}/{y}.png`).
-    /// Empty string means this source isn't proxied — the frontend's `url`
-    /// is expected to be an absolute upstream URL the browser hits directly
-    /// (the OSM case), and the tile-proxy route will refuse to serve it.
-    ///
-    /// Kept distinct from `url` because the two represent different layers:
-    /// `url` is what MapLibre fetches, `upstream_url` is what `reqwest` fetches
-    /// on the server. Conflating them is what caused PR #955.
-    #[serde(default)]
-    pub upstream_url: String,
-    /// Tile edge length in pixels. 256 for classic OSM-style sources.
-    #[serde(default = "default_tile_size")]
-    pub tile_size: u32,
-    /// Minimum zoom level the source serves tiles for.
-    #[serde(default)]
-    pub minzoom: u32,
-    /// Maximum zoom level the source serves tiles for.
-    #[serde(default = "default_tile_maxzoom")]
-    pub maxzoom: u32,
-    /// Attribution text shown in the MapLibre attribution control.
-    #[serde(default)]
-    pub attribution: String,
-    /// MapLibre `raster-saturation` paint (-1.0 to 1.0). Negative values
-    /// desaturate; 0.0 leaves colours untouched.
-    #[serde(default = "default_raster_saturation")]
-    pub raster_saturation: f32,
-    /// MapLibre `raster-opacity` paint (0.0 to 1.0).
-    #[serde(default = "default_raster_opacity")]
-    pub raster_opacity: f32,
-    /// When true, the frontend sets `scheme: 'tms'` on the MapLibre source,
-    /// flipping the y-axis for ArcGIS-style tile services.
-    #[serde(default)]
-    pub tms: bool,
-}
-
-fn default_tile_size() -> u32 {
-    256
-}
-fn default_tile_maxzoom() -> u32 {
-    19
-}
-fn default_raster_saturation() -> f32 {
-    -0.4
-}
-fn default_raster_opacity() -> f32 {
-    0.85
-}
-
-impl MapConfig {
-    /// Returns tile-source entries as `(id, label)` pairs, alphabetical by id.
-    /// Used by backends to populate [`parish_core::ipc::GameConfig::tile_sources`]
-    /// so the `/map` command handler can list and validate without needing
-    /// a reference to the whole engine config.
-    pub fn id_label_pairs(&self) -> Vec<(String, String)> {
-        self.tile_sources
-            .iter()
-            .map(|(id, src)| (id.clone(), src.label.clone()))
-            .collect()
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_engine_config_default() {
-        let cfg = EngineConfig::default();
-        assert_eq!(cfg.inference.timeout_secs, 300);
-        assert_eq!(cfg.inference.streaming_timeout_secs, 300);
-        assert_eq!(cfg.inference.log_capacity, 50);
-        assert!((cfg.speeds.normal - 36.0).abs() < f64::EPSILON);
-        assert!((cfg.encounters.dawn - 0.25).abs() < f64::EPSILON);
-        assert_eq!(cfg.npc.memory_capacity, 20);
-        assert!((cfg.palette.min_fg_bg_contrast - 80.0).abs() < f32::EPSILON);
-        assert!((cfg.world.fuzzy_threshold - 0.82).abs() < f64::EPSILON);
-        // PersistenceConfig is intentionally empty (TD-011)
-        let _ = cfg.persistence;
-    }
-
-    #[test]
-    fn test_engine_config_deserialize_empty() {
-        let cfg: EngineConfig = toml::from_str("").unwrap();
-        assert_eq!(cfg.inference.timeout_secs, 300);
-        assert_eq!(cfg.npc.memory_capacity, 20);
-    }
-
-    #[test]
-    fn test_engine_config_deserialize_partial() {
-        let toml_str = r#"
-[inference]
-timeout_secs = 60
-
-[npc]
-memory_capacity = 30
-"#;
-        let cfg: EngineConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(cfg.inference.timeout_secs, 60);
-        assert_eq!(cfg.inference.streaming_timeout_secs, 300); // default
-        assert_eq!(cfg.npc.memory_capacity, 30);
-        assert_eq!(cfg.npc.separator_holdback, 24); // default
-    }
-
-    #[test]
-    fn test_speed_config_defaults() {
-        let cfg = SpeedConfig::default();
-        assert!((cfg.slow - 18.0).abs() < f64::EPSILON);
-        assert!((cfg.normal - 36.0).abs() < f64::EPSILON);
-        assert!((cfg.fast - 72.0).abs() < f64::EPSILON);
-        assert!((cfg.fastest - 144.0).abs() < f64::EPSILON);
-        assert!((cfg.ludicrous - 864.0).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn test_encounter_config_defaults() {
-        let cfg = EncounterConfig::default();
-        assert!((cfg.dawn - 0.25).abs() < f64::EPSILON);
-        assert!((cfg.midnight - 0.05).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn test_npc_config_defaults() {
-        let cfg = NpcConfig::default();
-        assert_eq!(cfg.memory_capacity, 20);
-        assert_eq!(cfg.separator_holdback, 24);
-        assert_eq!(cfg.memory_context_count, 5);
-        assert_eq!(cfg.cognitive_tiers.tier1_max_distance, 0);
-        assert_eq!(cfg.cognitive_tiers.tier2_max_distance, 2);
-        assert!((cfg.relationship_labels.very_close - 0.7).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn test_palette_config_defaults() {
-        let cfg = PaletteConfig::default();
-        assert!((cfg.min_fg_bg_contrast - 80.0).abs() < f32::EPSILON);
-        assert!((cfg.min_muted_bg_contrast - 45.0).abs() < f32::EPSILON);
-    }
-
-    #[test]
-    fn test_world_config_defaults() {
-        let cfg = WorldConfig::default();
-        assert!((cfg.fuzzy_threshold - 0.82).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn test_persistence_config_default() {
-        let cfg = PersistenceConfig::default();
-        // Intentionally empty struct (TD-011); just verify it constructs.
-        let _ = cfg;
-    }
-
-    #[test]
-    fn test_inference_log_capacity_default() {
-        let cfg = InferenceConfig::default();
-        assert_eq!(cfg.log_capacity, 50);
-    }
-
-    #[test]
-    fn test_rate_limit_config_default_is_unset() {
-        let cfg = RateLimitConfig::default();
-        assert!(cfg.default.is_none());
-        assert!(cfg.dialogue.is_none());
-        assert!(cfg.simulation.is_none());
-        assert!(cfg.intent.is_none());
-        assert!(cfg.reaction.is_none());
-    }
-
-    #[test]
-    fn test_inference_config_default_has_no_rate_limits() {
-        let cfg = InferenceConfig::default();
-        assert!(cfg.rate_limits.default.is_none());
-        assert!(cfg.rate_limits.dialogue.is_none());
-    }
-
-    #[test]
-    fn test_rate_limit_config_for_category_returns_override() {
-        let cfg = RateLimitConfig {
-            dialogue: Some(CategoryRateLimit {
-                per_minute: 20,
-                burst: 4,
-            }),
-            simulation: Some(CategoryRateLimit {
-                per_minute: 60,
-                burst: 10,
-            }),
-            ..RateLimitConfig::default()
-        };
-        let dial = cfg.for_category(InferenceCategory::Dialogue).unwrap();
-        assert_eq!(dial.per_minute, 20);
-        assert_eq!(dial.burst, 4);
-        let sim = cfg.for_category(InferenceCategory::Simulation).unwrap();
-        assert_eq!(sim.per_minute, 60);
-        assert!(cfg.for_category(InferenceCategory::Intent).is_none());
-        assert!(cfg.for_category(InferenceCategory::Reaction).is_none());
-    }
-
-    #[test]
-    fn test_rate_limit_config_for_category_does_not_inherit_default() {
-        // The `default` field is for the base client, not per-category fallback.
-        let cfg = RateLimitConfig {
-            default: Some(CategoryRateLimit {
-                per_minute: 100,
-                burst: 5,
-            }),
-            ..RateLimitConfig::default()
-        };
-        assert!(cfg.for_category(InferenceCategory::Dialogue).is_none());
-    }
-
-    #[test]
-    fn test_category_rate_limit_burst_defaults_to_one() {
-        let toml = "per_minute = 30";
-        let cfg: CategoryRateLimit = toml::from_str(toml).unwrap();
-        assert_eq!(cfg.per_minute, 30);
-        assert_eq!(cfg.burst, 1);
-    }
-
-    #[test]
-    fn test_map_config_default_has_both_sources() {
-        let cfg = MapConfig::default();
-        assert_eq!(cfg.default_tile_source, "historic");
-        assert!(cfg.tile_sources.contains_key("osm"));
-        assert!(cfg.tile_sources.contains_key("historic"));
-        let osm = &cfg.tile_sources["osm"];
-        assert_eq!(osm.url, "https://tile.openstreetmap.org/{z}/{x}/{y}.png");
-        assert!(
-            osm.upstream_url.is_empty(),
-            "OSM is fetched directly by the browser; no server-side proxying"
-        );
-        assert_eq!(osm.tile_size, 256);
-        assert_eq!(osm.maxzoom, 19);
-        assert!(!osm.tms);
-        let historic = &cfg.tile_sources["historic"];
-        assert!(!historic.tms, "NLS serves standard XYZ, not TMS");
-        // Since issue #360, tiles are proxied through the local server so the
-        // client never hits NLS S3 directly.  `url` is the same-origin proxy
-        // path the browser hits; `upstream_url` is the absolute NLS S3 URL
-        // the server-side cache fetches from on a miss (PR #955).
-        assert!(
-            historic.url.starts_with("/tiles/historic/"),
-            "Historic 6\" tiles are proxied through the local server under the registered \
-             tile source id (issue #360); got: {}",
-            historic.url
-        );
-        assert!(
-            historic
-                .upstream_url
-                .starts_with("https://mapseries-tilesets.s3.amazonaws.com/os/roscommon1/"),
-            "Historic 6\" upstream_url must point at the NLS Roscommon 1st-edition \
-             S3 path so the server-side cache can fetch tiles on a miss; got: {}",
-            historic.upstream_url
-        );
-        assert_eq!(historic.maxzoom, 17, "NLS serves 6-inch up to z=17");
-    }
-
-    /// Pins the invariant that broke in PR #955: when a tile source's `url` is
-    /// a same-origin proxy path (`/tiles/<seg>/{z}/{x}/{y}.png`), its first
-    /// path segment must match the registering `tile_sources` key, because
-    /// `parish-server`'s tile-proxy route validates that segment against the
-    /// registered ids and 404s on mismatch.
-    #[test]
-    fn proxy_path_segment_matches_registered_source_id() {
-        let cfg = MapConfig::default();
-        for (id, src) in &cfg.tile_sources {
-            let Some(rest) = src.url.strip_prefix("/tiles/") else {
-                continue;
-            };
-            let seg = rest.split('/').next().unwrap_or("");
-            assert_eq!(
-                seg, id,
-                "tile source {id:?} should use its own id as the proxy path segment, \
-                 not {seg:?}; otherwise the tile-proxy route handler would 404 every \
-                 request and the cache lookup would happen under the wrong key"
-            );
-        }
-    }
-
-    #[test]
-    fn test_engine_config_includes_map_defaults() {
-        let cfg = EngineConfig::default();
-        assert_eq!(cfg.map.default_tile_source, "historic");
-        assert_eq!(cfg.map.tile_sources.len(), 2);
-    }
-
-    #[test]
-    fn test_map_config_apply_defaults_merges_missing_sources() {
-        let mut cfg = MapConfig {
-            default_tile_source: "osm".to_string(),
-            tile_sources: BTreeMap::new(),
-            bundled_tiles_dir: None,
-        };
-        cfg.tile_sources.insert(
-            "custom".to_string(),
-            TileSourceConfig {
-                label: "Custom".to_string(),
-                url: "https://example.com/{z}/{x}/{y}.png".to_string(),
-                upstream_url: String::new(),
-                tile_size: 256,
-                minzoom: 0,
-                maxzoom: 18,
-                attribution: "custom".to_string(),
-                raster_saturation: 0.0,
-                raster_opacity: 1.0,
-                tms: false,
-            },
-        );
-        cfg.apply_defaults();
-        assert!(cfg.tile_sources.contains_key("custom"));
-        assert!(cfg.tile_sources.contains_key("osm"));
-        assert!(cfg.tile_sources.contains_key("historic"));
-    }
-
-    #[test]
-    fn test_map_config_apply_defaults_preserves_user_overrides() {
-        let mut cfg = MapConfig::default();
-        cfg.tile_sources.get_mut("osm").unwrap().url =
-            "https://example.com/custom-osm/{z}/{x}/{y}.png".to_string();
-        cfg.apply_defaults();
-        assert_eq!(
-            cfg.tile_sources["osm"].url,
-            "https://example.com/custom-osm/{z}/{x}/{y}.png"
-        );
-    }
-
-    #[test]
-    fn map_config_bundled_tiles_dir_defaults_to_none() {
-        assert!(MapConfig::default().bundled_tiles_dir.is_none());
-    }
-
-    #[test]
-    fn map_config_bundled_tiles_dir_deserializes_from_toml() {
-        let toml = r#"bundled_tiles_dir = "/opt/parish/tiles""#;
-        let cfg: MapConfig = toml::from_str(toml).unwrap();
-        assert_eq!(
-            cfg.bundled_tiles_dir,
-            Some(std::path::PathBuf::from("/opt/parish/tiles"))
-        );
-    }
-
-    #[test]
-    fn map_config_apply_defaults_preserves_bundled_tiles_dir() {
-        let mut cfg = MapConfig {
-            bundled_tiles_dir: Some(std::path::PathBuf::from("/opt/tiles")),
-            ..MapConfig::default()
-        };
-        cfg.apply_defaults();
-        assert_eq!(
-            cfg.bundled_tiles_dir,
-            Some(std::path::PathBuf::from("/opt/tiles"))
-        );
-    }
-
-    #[test]
-    fn test_load_engine_config_missing_file() {
-        let cfg = load_engine_config(Path::new("/nonexistent/parish.toml"));
-        assert_eq!(cfg.map.default_tile_source, "historic");
-        assert_eq!(cfg.map.tile_sources.len(), 2);
-    }
-
-    #[test]
-    fn test_load_engine_config_from_file() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("parish.toml");
-        std::fs::write(
-            &path,
-            r#"
-[engine.map]
-default_tile_source = "historic"
-
-[engine.map.tile_sources.osm]
-url = "https://override/{z}/{x}/{y}.png"
-"#,
-        )
-        .unwrap();
-        let cfg = load_engine_config(&path);
-        assert_eq!(cfg.map.default_tile_source, "historic");
-        assert_eq!(
-            cfg.map.tile_sources.len(),
-            2,
-            "apply_defaults folded historic back in"
-        );
-        assert_eq!(
-            cfg.map.tile_sources["osm"].url,
-            "https://override/{z}/{x}/{y}.png"
-        );
-    }
-
-    #[test]
-    fn test_map_config_id_label_pairs_is_sorted() {
-        let cfg = MapConfig::default();
-        let pairs = cfg.id_label_pairs();
-        assert_eq!(pairs.len(), 2);
-        // BTreeMap iterates in sorted order, so "historic" < "osm".
-        assert_eq!(pairs[0].0, "historic");
-        assert_eq!(pairs[1].0, "osm");
-    }
-
-    #[test]
-    fn test_map_config_deserialize_partial_toml() {
-        // A partial override: user only overrides OSM URL, historic entry
-        // would be wiped without apply_defaults.
-        let toml_str = r#"
-[tile_sources.osm]
-url = "https://custom/{z}/{x}/{y}.png"
-attribution = "custom attribution"
-"#;
-        let mut cfg: MapConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(cfg.tile_sources.len(), 1, "serde replace semantics");
-        cfg.apply_defaults();
-        assert_eq!(
-            cfg.tile_sources.len(),
-            2,
-            "apply_defaults folds in historic"
-        );
-        assert_eq!(
-            cfg.tile_sources["osm"].url,
-            "https://custom/{z}/{x}/{y}.png"
-        );
-        assert_eq!(cfg.tile_sources["osm"].attribution, "custom attribution");
-    }
-
-    #[test]
-    fn test_inference_config_parses_rate_limits_from_toml() {
-        let toml_text = r#"
-            [rate_limits.default]
-            per_minute = 60
-            burst = 10
-
-            [rate_limits.dialogue]
-            per_minute = 20
-            burst = 4
-
-            [rate_limits.simulation]
-            per_minute = 30
-        "#;
-        let cfg: InferenceConfig = toml::from_str(toml_text).unwrap();
-        let default = cfg.rate_limits.default.unwrap();
-        assert_eq!(default.per_minute, 60);
-        assert_eq!(default.burst, 10);
-        let dial = cfg.rate_limits.dialogue.unwrap();
-        assert_eq!(dial.per_minute, 20);
-        assert_eq!(dial.burst, 4);
-        let sim = cfg.rate_limits.simulation.unwrap();
-        assert_eq!(sim.per_minute, 30);
-        assert_eq!(sim.burst, 1);
-        // Unspecified categories remain None
-        assert!(cfg.rate_limits.intent.is_none());
-        assert!(cfg.rate_limits.reaction.is_none());
-    }
-
-    #[test]
-    fn test_inference_config_parses_force_model_redownload_from_toml() {
-        let cfg: InferenceConfig = toml::from_str("force_model_redownload = true").unwrap();
-
-        assert!(cfg.force_model_redownload);
-        assert_eq!(cfg.model_download_timeout_secs, 3600);
-    }
-
-    #[test]
-    fn test_session_config_deserialize_from_toml() {
-        let toml_str = r#"
-idle_banter_after_secs = 60
-auto_pause_after_secs = 600
-max_concurrent_sessions = 10
-"#;
-        let cfg: SessionConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(cfg.idle_banter_after_secs, 60);
-        assert_eq!(cfg.auto_pause_after_secs, 600);
-        assert_eq!(cfg.max_concurrent_sessions, 10);
-    }
-
-    #[test]
-    fn test_session_config_deserialize_partial() {
-        let toml_str = "idle_banter_after_secs = 60";
-        let cfg: SessionConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(cfg.idle_banter_after_secs, 60);
-        assert_eq!(cfg.auto_pause_after_secs, 300);
-        assert_eq!(cfg.max_concurrent_sessions, 50);
-    }
-
-    #[test]
-    fn test_cognitive_tier_config_deserialize_from_toml() {
-        let toml_str = r#"
-tier1_max_distance = 1
-tier2_max_distance = 3
-tier3_max_distance = 10
-tier2_tick_interval_minutes = 10
-tier3_tick_interval_hours = 12
-tier3_batch_size = 20
-tier4_tick_interval_days = 30
-"#;
-        let cfg: CognitiveTierConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(cfg.tier1_max_distance, 1);
-        assert_eq!(cfg.tier2_max_distance, 3);
-        assert_eq!(cfg.tier3_max_distance, 10);
-        assert_eq!(cfg.tier2_tick_interval_minutes, 10);
-        assert_eq!(cfg.tier3_tick_interval_hours, 12);
-        assert_eq!(cfg.tier3_batch_size, 20);
-        assert_eq!(cfg.tier4_tick_interval_days, 30);
-    }
-
-    #[test]
-    fn test_cognitive_tier_config_deserialize_partial() {
-        let toml_str = "tier1_max_distance = 1";
-        let cfg: CognitiveTierConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(cfg.tier1_max_distance, 1);
-        assert_eq!(cfg.tier2_max_distance, 2);
-    }
-
-    #[test]
-    fn test_relationship_label_config_deserialize_from_toml() {
-        let toml_str = r#"
-very_close = 0.8
-friendly = 0.4
-acquainted = 0.1
-cool = -0.2
-strained = -0.6
-"#;
-        let cfg: RelationshipLabelConfig = toml::from_str(toml_str).unwrap();
-        assert!((cfg.very_close - 0.8).abs() < f64::EPSILON);
-        assert!((cfg.friendly - 0.4).abs() < f64::EPSILON);
-        assert!((cfg.acquainted - 0.1).abs() < f64::EPSILON);
-        assert!((cfg.cool - (-0.2)).abs() < f64::EPSILON);
-        assert!((cfg.strained - (-0.6)).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn test_reaction_config_deserialize_from_toml() {
-        let toml_str = r#"
-base_chance = 0.8
-workplace_bonus = 0.2
-indoor_bonus = 0.05
-empathy_bonus = 0.1
-negative_mood_penalty = 0.1
-night_penalty = 0.05
-llm_timeout_secs = 10
-max_reactions = 5
-"#;
-        let cfg: ReactionConfig = toml::from_str(toml_str).unwrap();
-        assert!((cfg.base_chance - 0.8).abs() < f64::EPSILON);
-        assert!((cfg.workplace_bonus - 0.2).abs() < f64::EPSILON);
-        assert!((cfg.indoor_bonus - 0.05).abs() < f64::EPSILON);
-        assert!((cfg.empathy_bonus - 0.1).abs() < f64::EPSILON);
-        assert!((cfg.negative_mood_penalty - 0.10).abs() < f64::EPSILON);
-        assert!((cfg.night_penalty - 0.05).abs() < f64::EPSILON);
-        assert_eq!(cfg.llm_timeout_secs, 10);
-        assert_eq!(cfg.max_reactions, 5);
-    }
-
-    #[test]
-    fn test_encounter_config_deserialize_from_toml() {
-        let toml_str = r#"
-dawn = 0.30
-morning = 0.25
-midday = 0.20
-afternoon = 0.15
-dusk = 0.10
-night = 0.05
-midnight = 0.02
-"#;
-        let cfg: EncounterConfig = toml::from_str(toml_str).unwrap();
-        assert!((cfg.dawn - 0.30).abs() < f64::EPSILON);
-        assert!((cfg.morning - 0.25).abs() < f64::EPSILON);
-        assert!((cfg.midday - 0.20).abs() < f64::EPSILON);
-        assert!((cfg.afternoon - 0.15).abs() < f64::EPSILON);
-        assert!((cfg.dusk - 0.10).abs() < f64::EPSILON);
-        assert!((cfg.night - 0.05).abs() < f64::EPSILON);
-        assert!((cfg.midnight - 0.02).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn test_encounter_config_deserialize_partial() {
-        let toml_str = "dawn = 0.10";
-        let cfg: EncounterConfig = toml::from_str(toml_str).unwrap();
-        assert!((cfg.dawn - 0.10).abs() < f64::EPSILON);
-        assert!((cfg.morning - 0.25).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn test_palette_config_deserialize_from_toml() {
-        let toml_str = r#"
-min_fg_bg_contrast = 90.0
-min_muted_bg_contrast = 50.0
-"#;
-        let cfg: PaletteConfig = toml::from_str(toml_str).unwrap();
-        assert!((cfg.min_fg_bg_contrast - 90.0).abs() < f32::EPSILON);
-        assert!((cfg.min_muted_bg_contrast - 50.0).abs() < f32::EPSILON);
-    }
-
-    #[test]
-    fn test_palette_config_deserialize_partial() {
-        let toml_str = "min_fg_bg_contrast = 70.0";
-        let cfg: PaletteConfig = toml::from_str(toml_str).unwrap();
-        assert!((cfg.min_fg_bg_contrast - 70.0).abs() < f32::EPSILON);
-        assert!((cfg.min_muted_bg_contrast - 45.0).abs() < f32::EPSILON);
-    }
-
-    #[test]
-    fn test_world_config_deserialize_from_toml() {
-        let toml_str = "fuzzy_threshold = 0.90";
-        let cfg: WorldConfig = toml::from_str(toml_str).unwrap();
-        assert!((cfg.fuzzy_threshold - 0.90).abs() < f64::EPSILON);
-    }
-
-    #[test]
-    fn test_persistence_config_deserialize_from_toml() {
-        let cfg: PersistenceConfig = toml::from_str("").unwrap();
-        // Intentionally empty struct (TD-011); just verify it parses.
-        let _ = cfg;
-    }
-
-    #[test]
-    fn test_inference_config_deserialize_from_toml() {
-        let toml_str = r#"
-timeout_secs = 45
-streaming_timeout_secs = 600
-reachability_timeout_secs = 15
-model_download_timeout_secs = 7200
-force_model_redownload = true
-model_loading_timeout_secs = 600
-log_capacity = 100
-"#;
-        let cfg: InferenceConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(cfg.timeout_secs, 45);
-        assert_eq!(cfg.streaming_timeout_secs, 600);
-        assert_eq!(cfg.reachability_timeout_secs, 15);
-        assert_eq!(cfg.model_download_timeout_secs, 7200);
-        assert!(cfg.force_model_redownload);
-        assert_eq!(cfg.model_loading_timeout_secs, 600);
-        assert_eq!(cfg.log_capacity, 100);
-        assert!(cfg.rate_limits.default.is_none());
-    }
-
-    #[test]
-    fn test_inference_config_deserialize_partial() {
-        let toml_str = "timeout_secs = 60";
-        let cfg: InferenceConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(cfg.timeout_secs, 60);
-        assert_eq!(cfg.streaming_timeout_secs, 300);
-    }
-
-    #[test]
-    fn test_map_config_deserialize_from_toml() {
-        let toml_str = r#"
-default_tile_source = "osm"
-
-[tile_sources.osm]
-label = "OpenStreetMap"
-url = "https://tile.openstreetmap.org/{z}/{x}/{y}.png"
-tile_size = 256
-minzoom = 0
-maxzoom = 19
-attribution = "© OpenStreetMap contributors"
-raster_saturation = -0.4
-raster_opacity = 0.85
-tms = false
-"#;
-        let cfg: MapConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(cfg.default_tile_source, "osm");
-        assert_eq!(cfg.tile_sources.len(), 1);
-        let osm = &cfg.tile_sources["osm"];
-        assert_eq!(osm.label, "OpenStreetMap");
-        assert_eq!(osm.url, "https://tile.openstreetmap.org/{z}/{x}/{y}.png");
-        assert_eq!(osm.tile_size, 256);
-        assert_eq!(osm.minzoom, 0);
-        assert_eq!(osm.maxzoom, 19);
-        assert_eq!(osm.attribution, "© OpenStreetMap contributors");
-        assert!((osm.raster_saturation - (-0.4)).abs() < f32::EPSILON);
-        assert!((osm.raster_opacity - 0.85).abs() < f32::EPSILON);
-        assert!(!osm.tms);
-    }
-
-    // ── #1224 — dialogue display cap (AC-5) ──────────────────────────────────
-
-    /// AC-5 (fix-1224-1225): `NpcConfig` must expose `dialogue_display_max_chars`
-    /// with a sensible default, and it must survive round-trip TOML serialisation
-    /// (old configs that omit the field must deserialise to the default).
-    #[test]
-    fn npc_config_has_dialogue_display_max_chars_with_sensible_default() {
-        let cfg = NpcConfig::default();
-        // Default should be > 0 (cap enabled) and > typical 2-4 sentence reply.
-        assert!(
-            cfg.dialogue_display_max_chars > 0,
-            "cap must be enabled by default (non-zero)"
-        );
-        assert!(
-            cfg.dialogue_display_max_chars >= 500,
-            "cap must be generous enough not to clip normal 2-4 sentence replies"
-        );
-    }
-
-    /// Old config without `dialogue_display_max_chars` must deserialise cleanly,
-    /// falling back to the default value (serde `default` attribute).
-    #[test]
-    fn npc_config_dialogue_display_max_chars_defaults_when_absent() {
-        let toml_str = r#"
-[npc]
-memory_capacity = 25
-"#;
-        let cfg: EngineConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(cfg.npc.memory_capacity, 25);
-        assert_eq!(
-            cfg.npc.dialogue_display_max_chars,
-            NpcConfig::default().dialogue_display_max_chars,
-            "missing field must deserialise to the default"
-        );
-    }
-
-    // ── #1228 — dialogue repetition threshold (AC-5) ─────────────────────────
-
-    /// AC-5 (fix-1228): `NpcConfig` exposes `dialogue_repetition_threshold` with
-    /// a sensible default that enables the cross-turn guard by default.
-    #[test]
-    fn npc_config_has_dialogue_repetition_threshold_enabled_by_default() {
-        let cfg = NpcConfig::default();
-        assert!(
-            cfg.dialogue_repetition_threshold > 0.0,
-            "cross-turn guard must be enabled by default (threshold > 0)"
-        );
-        assert!(
-            cfg.dialogue_repetition_threshold <= 1.0,
-            "threshold is a Jaccard similarity in [0, 1]"
-        );
-        assert!(
-            cfg.dialogue_repetition_threshold >= 0.8,
-            "threshold must be strict enough to avoid flagging normal variation"
-        );
-    }
-
-    /// Old config without `dialogue_repetition_threshold` must deserialise
-    /// cleanly, falling back to the default (serde `default`).
-    #[test]
-    fn npc_config_dialogue_repetition_threshold_defaults_when_absent() {
-        let toml_str = r#"
-[npc]
-memory_capacity = 25
-"#;
-        let cfg: EngineConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(
-            cfg.npc.dialogue_repetition_threshold,
-            NpcConfig::default().dialogue_repetition_threshold,
-            "missing field must deserialise to the default"
-        );
-    }
-}
