+++ parish/crates/parish-config/src/provider/mod.rs
@@ -16,878 +16,45 @@
 //! point the registry is read-only in practice; existing
 //! `Arc<ProviderMod>` handles keep pointing at their snapshot regardless
 //! of subsequent merges.
-
-use crate::builtin_providers;
-use parish_types::ParishError;
-use serde::Deserialize;
-use std::collections::HashMap;
-use std::path::Path;
-use std::sync::{Arc, OnceLock, RwLock};
-
-// ── ProviderKind ─────────────────────────────────────────────────────────────
-
-/// Client-routing category. Controls which HTTP client (Anthropic
-/// Messages vs OpenAI-compat vs Simulator) is used.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, serde::Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum ProviderKind {
-    Anthropic,
-    #[serde(rename = "openai-compat")]
-    OpenAiCompat,
-    /// Ollama, LM Studio, vLLM — OpenAI-compat on the wire but managed locally.
-    Local,
-    Simulator,
-}
-
-// ── ProviderPreset ────────────────────────────────────────────────────────────
-
-/// Per-category base-URL overrides shipped with a provider preset.
-///
-/// Two-slot loadouts (e.g. vllm-mlx on Apple Silicon: 14B on :8000 + 1.5B
-/// on :8001) need each category routed to the slot where its preset model
-/// is actually loaded. Without this, `fill_missing_models_from_presets`
-/// picks the preset model but inherits the base URL — guaranteeing a
-/// model/URL mismatch and a 404 on every reaction/simulation call.
-#[derive(Debug, Clone, Default, PartialEq, Eq, serde::Deserialize)]
-pub struct PresetBaseUrls {
-    pub dialogue: Option<String>,
-    pub simulation: Option<String>,
-    pub intent: Option<String>,
-    pub reaction: Option<String>,
-}
-
-impl PresetBaseUrls {
-    pub fn url(&self, cat: InferenceCategory) -> Option<&str> {
-        match cat {
-            InferenceCategory::Dialogue => self.dialogue.as_deref(),
-            InferenceCategory::Simulation => self.simulation.as_deref(),
-            InferenceCategory::Intent => self.intent.as_deref(),
-            InferenceCategory::Reaction => self.reaction.as_deref(),
-        }
-    }
-}
-
-/// One named model configuration shipped with a provider.
-#[derive(Debug, Clone, PartialEq, Eq, serde::Deserialize)]
-pub struct ProviderPreset {
-    pub key: String,
-    pub label: String,
-    pub dialogue: Option<String>,
-    pub simulation: Option<String>,
-    pub intent: Option<String>,
-    pub reaction: Option<String>,
-    /// Per-category base URL hints. When unset for a category the user's
-    /// base URL is inherited (single-slot providers). See [`PresetBaseUrls`].
-    #[serde(default)]
-    pub base_urls: PresetBaseUrls,
-}
-
-impl ProviderPreset {
-    pub fn model(&self, cat: InferenceCategory) -> Option<&str> {
-        match cat {
-            InferenceCategory::Dialogue => self.dialogue.as_deref(),
-            InferenceCategory::Simulation => self.simulation.as_deref(),
-            InferenceCategory::Intent => self.intent.as_deref(),
-            InferenceCategory::Reaction => self.reaction.as_deref(),
-        }
-    }
-
-    pub fn base_url(&self, cat: InferenceCategory) -> Option<&str> {
-        self.base_urls.url(cat)
-    }
-}
-
-// ── ProviderMod ───────────────────────────────────────────────────────────────
-
-/// All data about one provider, deserialized from TOML.
-#[derive(Debug, Clone, PartialEq, Eq, serde::Deserialize)]
-pub struct ProviderMod {
-    pub id: String,
-    pub display_name: String,
-    #[serde(default)]
-    pub aliases: Vec<String>,
-    pub kind: ProviderKind,
-    pub default_base_url: String,
-    #[serde(default)]
-    pub requires_api_key: bool,
-    #[serde(default)]
-    pub needs_base_url_from_user: bool,
-    #[serde(default = "default_true")]
-    pub requires_model: bool,
-    pub api_key_env_var: Option<String>,
-    pub blurb: Option<String>,
-    pub signup_url: Option<String>,
-    #[serde(default)]
-    pub featured: bool,
-    /// True when the provider is local inference where API keys are
-    /// irrelevant (ollama, lmstudio, vllm, vllm_mlx, simulator). The
-    /// onboarding wizard uses this to relax model-name/key guards.
-    /// Distinct from `requires_api_key`: `custom` does not require an
-    /// API key but still needs a model name and base URL, so its
-    /// `keyless` is `false` (codex P2 regression fix).
-    #[serde(default)]
-    pub keyless: bool,
-    #[serde(default)]
-    pub presets: Vec<ProviderPreset>,
-}
-
-fn default_true() -> bool {
-    true
-}
-
-impl ProviderMod {
-    pub fn has_preset(&self) -> bool {
-        !self.presets.is_empty()
-    }
-
-    pub fn preset_model(&self, cat: InferenceCategory) -> Option<&str> {
-        self.presets.first()?.model(cat)
-    }
-
-    /// Per-category base URL from the recommended preset, when supplied.
-    /// Returns `None` if the preset omits the field (single-slot providers).
-    pub fn preset_base_url(&self, cat: InferenceCategory) -> Option<&str> {
-        self.presets.first()?.base_url(cat)
-    }
-
-    pub fn preset_models_array(&self) -> [Option<&str>; 4] {
-        let first = self.presets.first();
-        [
-            first.and_then(|p| p.dialogue.as_deref()),
-            first.and_then(|p| p.simulation.as_deref()),
-            first.and_then(|p| p.intent.as_deref()),
-            first.and_then(|p| p.reaction.as_deref()),
-        ]
-    }
-}
-
-// ── Provider ──────────────────────────────────────────────────────────────────
-
-/// Returns the host's total physical memory in bytes, or `None` on
-/// platforms where we can't read it cheaply.
-pub fn unified_memory_bytes() -> Option<u64> {
-    #[cfg(target_os = "macos")]
-    {
-        use std::process::Command;
-        let output = Command::new("sysctl")
-            .args(["-n", "hw.memsize"])
-            .output()
-            .ok()?;
-        if !output.status.success() {
-            return None;
-        }
-        let s = std::str::from_utf8(&output.stdout).ok()?;
-        s.trim().parse::<u64>().ok()
-    }
-    #[cfg(not(target_os = "macos"))]
-    {
-        None
-    }
-}
-
-/// A loaded provider configuration. Wraps `Arc<ProviderMod>` for
-/// cheap cloning and shared ownership.
-#[derive(Debug, Clone)]
-pub struct Provider(pub Arc<ProviderMod>);
-
-impl Provider {
-    pub fn id(&self) -> &str {
-        &self.0.id
-    }
-    pub fn kind(&self) -> ProviderKind {
-        self.0.kind
-    }
-    pub fn default_base_url(&self) -> &str {
-        &self.0.default_base_url
-    }
-    pub fn requires_api_key(&self) -> bool {
-        self.0.requires_api_key
-    }
-    pub fn needs_base_url_from_user(&self) -> bool {
-        self.0.needs_base_url_from_user
-    }
-    pub fn requires_model(&self) -> bool {
-        self.0.requires_model
-    }
-    pub fn api_key_env_var(&self) -> Option<&str> {
-        self.0.api_key_env_var.as_deref()
-    }
-    pub fn presets(&self) -> &[ProviderPreset] {
-        &self.0.presets
-    }
-    pub fn has_preset(&self) -> bool {
-        self.0.has_preset()
-    }
-    pub fn preset_model(&self, cat: InferenceCategory) -> Option<&str> {
-        self.0.preset_model(cat)
-    }
-    pub fn preset_base_url(&self, cat: InferenceCategory) -> Option<&str> {
-        self.0.preset_base_url(cat)
-    }
-    /// Returns `[dialogue, simulation, intent, reaction]` from the first preset.
-    pub fn preset_models(&self) -> [Option<&str>; 4] {
-        self.0.preset_models_array()
-    }
-    pub fn display_name(&self) -> &str {
-        &self.0.display_name
-    }
-
-    pub fn is_configured_in_env(&self) -> bool {
-        if !self.requires_api_key() {
-            return true;
-        }
-        self.api_key_env_var()
-            .and_then(|v| std::env::var(v).ok())
-            .map(|v| !v.is_empty())
-            .unwrap_or(false)
-    }
-
-    pub fn recommended_for_platform() -> Self {
-        if cfg!(target_os = "macos") {
-            if unified_memory_bytes().unwrap_or(0) >= 16 * 1_073_741_824 {
-                registry()
-                    .get("vllmmlx")
-                    .expect("vllmmlx must be registered")
-            } else {
-                registry()
-                    .get("simulator")
-                    .expect("simulator must be registered")
-            }
-        } else {
-            registry().get("vllm").expect("vllm must be registered")
-        }
-    }
-
-    pub fn from_id(id: &str) -> Option<Self> {
-        registry().get(id)
-    }
-
-    pub fn from_str_loose(s: &str) -> Result<Self, ParishError> {
-        registry().lookup(s)
-    }
-
-    // Builtin convenience constructors. Builtins are always present.
-    pub fn ollama() -> Self {
-        Self::from_id("ollama").expect("ollama builtin must be registered")
-    }
-    pub fn simulator() -> Self {
-        Self::from_id("simulator").expect("simulator builtin must be registered")
-    }
-    pub fn custom() -> Self {
-        Self::from_id("custom").expect("custom builtin must be registered")
-    }
-    pub fn vllm() -> Self {
-        Self::from_id("vllm").expect("vllm builtin must be registered")
-    }
-    pub fn vllmmlx() -> Self {
-        Self::from_id("vllmmlx").expect("vllmmlx builtin must be registered")
-    }
-}
-
-impl PartialEq for Provider {
-    fn eq(&self, other: &Self) -> bool {
-        self.0.id == other.0.id
-    }
-}
-impl Eq for Provider {}
-impl std::hash::Hash for Provider {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.0.id.hash(state);
-    }
-}
-impl Default for Provider {
-    fn default() -> Self {
-        Self::simulator()
-    }
-}
-
-impl std::fmt::Display for Provider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0.id)
-    }
-}
-
-// ── ProviderRegistry ─────────────────────────────────────────────────────────
-
-/// Provider registry. Uses interior mutability so the static `REGISTRY`
-/// can be pre-populated with builtins and then have mod-loaded providers
-/// merged in post-init via `register_mod_providers`. Existing
-/// `Arc<ProviderMod>` handles handed out before a merge keep pointing at
-/// their snapshot.
-pub struct ProviderRegistry {
-    by_id: RwLock<HashMap<String, Provider>>,
-}
-
-impl ProviderRegistry {
-    /// Loads the registry pre-populated with the five engine builtins.
-    /// Cloud providers are *not* loaded here; they arrive later via
-    /// `register_mod_providers` after `discover_mods` runs in bootstrap.
-    fn load_with_builtins() -> Self {
-        let mut by_id: HashMap<String, Provider> = HashMap::new();
-
-        for raw in builtin_providers::ALL {
-            let m: ProviderMod = toml::from_str(raw)
-                .expect("builtin provider TOML parse failed — check builtin_providers/*.toml");
-            let p = Provider(Arc::new(m));
-            by_id.insert(p.id().to_string(), p);
-        }
-
-        ProviderRegistry {
-            by_id: RwLock::new(by_id),
-        }
-    }
-
-    pub fn get(&self, id: &str) -> Option<Provider> {
-        self.by_id
-            .read()
-            .expect("registry poisoned")
-            .get(id)
-            .cloned()
-    }
-
-    pub fn lookup(&self, s: &str) -> Result<Provider, ParishError> {
-        let lower = s.to_lowercase();
-        let guard = self.by_id.read().expect("registry poisoned");
-        if let Some(p) = guard.get(&lower) {
-            return Ok(p.clone());
-        }
-        for p in guard.values() {
-            if p.0.aliases.iter().any(|a| a == &lower) {
-                return Ok(p.clone());
-            }
-        }
-        let mut known: Vec<&str> = guard.keys().map(String::as_str).collect();
-        known.sort();
-        Err(ParishError::Config(format!(
-            "unknown provider '{}'. Known: {}",
-            s,
-            known.join(", ")
-        )))
-    }
-
-    pub fn all(&self) -> Vec<Provider> {
-        let mut v: Vec<_> = self
-            .by_id
-            .read()
-            .expect("registry poisoned")
-            .values()
-            .cloned()
-            .collect();
-        v.sort_by(|a, b| a.id().cmp(b.id()));
-        v
-    }
-
-    pub fn featured(&self) -> Vec<Provider> {
-        let mut v: Vec<_> = self
-            .by_id
-            .read()
-            .expect("registry poisoned")
-            .values()
-            .filter(|p| p.0.featured)
-            .cloned()
-            .collect();
-        v.sort_by(|a, b| a.id().cmp(b.id()));
-        v
-    }
-
-    /// Merges runtime-loaded providers into the registry. Called from the
-    /// bootstrap path once per process, after `discover_mods` collects
-    /// `ModKind::Providers` mods. Last-wins on id collision with a WARN
-    /// log; the collision message is the only operator signal that two
-    /// mods (or a mod and a builtin) ship the same provider id.
-    pub fn register_mod_providers(&self, mods: Vec<ProviderMod>) -> Result<(), ParishError> {
-        let mut guard = self.by_id.write().expect("registry poisoned");
-        for m in mods {
-            let id = m.id.clone();
-            if let Some(existing) = guard.get(&id) {
-                // Silent no-op when re-registering identical content — the
-                // auto-loader and bootstrap may both fire in debug builds
-                // and that overlap should not log spam.
-                if existing.0.as_ref() == &m {
-                    continue;
-                }
-                tracing::warn!(
-                    "mod-loaded provider '{}' overrides existing registry entry (last-wins)",
-                    id
-                );
-            }
-            guard.insert(id, Provider(Arc::new(m)));
-        }
-        Ok(())
-    }
-}
-
-static REGISTRY: OnceLock<ProviderRegistry> = OnceLock::new();
-
-pub fn registry() -> &'static ProviderRegistry {
-    let r = REGISTRY.get_or_init(ProviderRegistry::load_with_builtins);
-    // Auto-discover and register provider mods on first access. This
-    // runs in all build profiles so that release startup paths which
-    // resolve provider config before the explicit bootstrap
-    // (`parish_core::game_mod::register_provider_mods_once`) still see
-    // the same registry as debug/test builds. The walk is idempotent
-    // (guarded by `Once`) and silently no-ops when no `mods/` directory
-    // is discoverable — production deployments that need a non-default
-    // location set `PARISH_MODS_DIR`.
-    ensure_mods_loaded();
-    r
-}
-
-/// Discover and register every `mods/<id>/providers/*.toml` from the
-/// first directory found via, in priority order:
-///
-/// 1. `PARISH_MODS_DIR` env var (operator override for packaged builds);
-/// 2. walk-up from this crate's compile-time `CARGO_MANIFEST_DIR` to a
-///    directory containing `mods/` (dev tree, `cargo test`/`cargo run`).
-///
-/// Idempotent — guarded by an internal `Once`; safe to call repeatedly.
-///
-/// Returns silently when no mods directory is discoverable. The explicit
-/// `parish_core::game_mod::register_provider_mods_once` bootstrap path is
-/// still authoritative for runtime mod loads with surfaced error
-/// reporting; this helper exists so that any `Provider::from_id` /
-/// `from_str_loose` call hit before that bootstrap (release startup,
-/// downstream test crates, env-var-only cloud configs) still finds the
-/// shipped provider mods.
-pub fn ensure_mods_loaded() {
-    use std::sync::Once;
-    static INIT: Once = Once::new();
-    INIT.call_once(|| {
-        let mods_dir = std::env::var_os("PARISH_MODS_DIR")
-            .map(std::path::PathBuf::from)
-            .filter(|p| p.is_dir())
-            .or_else(|| {
-                let crate_dir = std::path::Path::new(env!("CARGO_MANIFEST_DIR"));
-                let mut dir = crate_dir.to_path_buf();
-                loop {
-                    let candidate = dir.join("mods");
-                    if candidate.is_dir() {
-                        break Some(candidate);
-                    }
-                    if !dir.pop() {
-                        break None;
-                    }
-                }
-            });
-        let Some(mods_dir) = mods_dir else { return };
-        let Ok(read) = std::fs::read_dir(&mods_dir) else {
-            return;
-        };
-        let mut all: Vec<ProviderMod> = Vec::new();
-        for entry in read.flatten() {
-            let providers = entry.path().join("providers");
-            if !providers.is_dir() {
-                continue;
-            }
-            let Ok(files) = std::fs::read_dir(&providers) else {
-                continue;
-            };
-            for f in files.flatten() {
-                if f.path().extension().is_none_or(|x| x != "toml") {
-                    continue;
-                }
-                let Ok(raw) = std::fs::read_to_string(f.path()) else {
-                    continue;
-                };
-                if let Ok(m) = toml::from_str::<ProviderMod>(&raw) {
-                    all.push(m);
-                }
-            }
-        }
-        // Avoid recursing back through `registry()` — it would re-enter this
-        // `Once::call_once` and deadlock. Use the underlying static directly,
-        // which is guaranteed initialized by the caller in `registry()`.
-        let r = REGISTRY.get_or_init(ProviderRegistry::load_with_builtins);
-        let _ = r.register_mod_providers(all);
-    });
-}
-
-// ── InferenceCategory ────────────────────────────────────────────────────────
-
-/// Inference categories that can each have independent provider/model/key settings.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum InferenceCategory {
-    /// Player-facing NPC dialogue (Tier 1, streaming).
-    Dialogue,
-    /// Background NPC simulation (Tier 2, JSON).
-    Simulation,
-    /// Player input intent parsing (JSON, low-latency).
-    Intent,
-    /// NPC arrival reactions/greetings (short timeout, fast model).
-    Reaction,
-}
-
-impl InferenceCategory {
-    /// All defined inference categories.
-    pub const ALL: [InferenceCategory; 4] = [
-        InferenceCategory::Dialogue,
-        InferenceCategory::Simulation,
-        InferenceCategory::Intent,
-        InferenceCategory::Reaction,
-    ];
-
-    /// Array index matching [`InferenceCategory::ALL`] order.
-    pub fn idx(self) -> usize {
-        match self {
-            InferenceCategory::Dialogue => 0,
-            InferenceCategory::Simulation => 1,
-            InferenceCategory::Intent => 2,
-            InferenceCategory::Reaction => 3,
-        }
-    }
-
-    /// Returns the lowercase name used in TOML keys, env var prefixes, and CLI flags.
-    pub fn name(&self) -> &'static str {
-        match self {
-            InferenceCategory::Dialogue => "dialogue",
-            InferenceCategory::Simulation => "simulation",
-            InferenceCategory::Intent => "intent",
-            InferenceCategory::Reaction => "reaction",
-        }
-    }
-
-    /// Parses a category name (case-insensitive).
-    pub fn from_name(s: &str) -> Option<Self> {
-        match s.to_lowercase().as_str() {
-            "dialogue" => Some(InferenceCategory::Dialogue),
-            "simulation" => Some(InferenceCategory::Simulation),
-            "intent" => Some(InferenceCategory::Intent),
-            "reaction" => Some(InferenceCategory::Reaction),
-            _ => None,
-        }
-    }
-
-    /// Returns the SCREAMING_CASE prefix used in environment variables.
-    pub fn env_prefix(&self) -> &'static str {
-        match self {
-            InferenceCategory::Dialogue => "PARISH_DIALOGUE",
-            InferenceCategory::Simulation => "PARISH_SIMULATION",
-            InferenceCategory::Intent => "PARISH_INTENT",
-            InferenceCategory::Reaction => "PARISH_REACTION",
-        }
-    }
-}
-
-// ── ProviderConfig ────────────────────────────────────────────────────────────
-
-/// Resolved provider configuration ready for use.
-#[derive(Debug, Clone)]
-pub struct ProviderConfig {
-    /// The selected provider backend.
-    pub provider: Provider,
-    /// Base URL for the provider API.
-    pub base_url: String,
-    /// API key for authenticated providers (OpenRouter, etc.).
-    pub api_key: Option<String>,
-    /// Model name override. Required for non-Ollama providers.
-    pub model: Option<String>,
-}
-
-/// Resolved cloud provider configuration for player-facing dialogue.
-#[derive(Debug, Clone)]
-pub struct CloudConfig {
-    pub provider: Provider,
-    pub base_url: String,
-    pub api_key: Option<String>,
-    pub model: String,
-}
-
-/// CLI-provided overrides for cloud provider configuration.
-#[derive(Debug, Default)]
-pub struct CliCloudOverrides {
-    pub provider: Option<String>,
-    pub base_url: Option<String>,
-    pub model: Option<String>,
-}
-
-/// Raw TOML file structure for `parish.toml`.
-#[derive(Debug, Deserialize, Default)]
-struct TomlConfig {
-    #[serde(default)]
-    provider: TomlProvider,
-    #[serde(default)]
-    cloud: TomlCloud,
-}
-
-/// The `[provider]` section of the TOML config.
-#[derive(Debug, Deserialize, Default)]
-struct TomlProvider {
-    name: Option<String>,
-    base_url: Option<String>,
-    api_key: Option<String>,
-    model: Option<String>,
-}
-
-/// The `[cloud]` section of the TOML config for cloud dialogue provider.
-#[derive(Debug, Deserialize, Default)]
-struct TomlCloud {
-    name: Option<String>,
-    base_url: Option<String>,
-    api_key: Option<String>,
-    model: Option<String>,
-}
-
-/// CLI-provided overrides for provider configuration.
-#[derive(Debug, Default)]
-pub struct CliOverrides {
-    pub provider: Option<String>,
-    pub base_url: Option<String>,
-    pub model: Option<String>,
-}
-
-impl ProviderConfig {
-    /// Returns a display-friendly provider name.
-    pub fn provider_display(&self) -> String {
-        self.provider.id().to_string()
-    }
-}
-
-// ── Shared 4-layer resolution helper ─────────────────────────────────────────
-
-struct RawLayers {
-    provider_str: Option<String>,
-    base_url: Option<String>,
-    api_key: Option<String>,
-    model: Option<String>,
-}
-
-fn apply_env_and_cli_layers(
-    mut raw: RawLayers,
-    env_prefix: &str,
-    cli_provider: Option<&str>,
-    cli_base_url: Option<&str>,
-    cli_model: Option<&str>,
-) -> RawLayers {
-    if let Some(val) = env_non_empty(&format!("{env_prefix}_PROVIDER")) {
-        raw.provider_str = Some(val);
-    }
-    if let Some(val) = env_non_empty(&format!("{env_prefix}_BASE_URL")) {
-        raw.base_url = Some(val);
-    }
-    if let Some(val) = env_non_empty(&format!("{env_prefix}_MODEL")) {
-        raw.model = Some(val);
-    }
-
-    if let Some(val) = cli_provider {
-        raw.provider_str = Some(val.to_string());
-    }
-    if let Some(val) = cli_base_url {
-        raw.base_url = Some(val.to_string());
-    }
-    if let Some(val) = cli_model {
-        raw.model = Some(val.to_string());
-    }
-
-    raw
-}
-
-fn load_toml(config_path: Option<&Path>) -> Result<TomlConfig, ParishError> {
-    match config_path {
-        Some(path) => read_toml_config(path),
-        None => Ok(TomlConfig::default()),
-    }
-}
-
-/// Resolves provider configuration from file, env vars, and CLI flags.
-pub fn resolve_config(
-    config_path: Option<&Path>,
-    cli: &CliOverrides,
-) -> Result<ProviderConfig, ParishError> {
-    let toml_cfg = load_toml(config_path)?;
-
-    let toml_raw = RawLayers {
-        provider_str: toml_cfg.provider.name,
-        base_url: toml_cfg.provider.base_url,
-        api_key: toml_cfg.provider.api_key,
-        model: toml_cfg.provider.model,
-    };
-    let mut raw = apply_env_and_cli_layers(
-        toml_raw,
-        "PARISH",
-        cli.provider.as_deref(),
-        cli.base_url.as_deref(),
-        cli.model.as_deref(),
-    );
-
-    // Deprecated PARISH_OLLAMA_URL fallback
-    if raw.base_url.is_none()
-        && let Some(val) = env_non_empty("PARISH_OLLAMA_URL")
-    {
-        tracing::warn!("PARISH_OLLAMA_URL is deprecated, use PARISH_BASE_URL instead");
-        raw.base_url = Some(val);
-    }
-
-    let provider = match &raw.provider_str {
-        Some(s) => Provider::from_str_loose(s)?,
-        None => Provider::default(),
-    };
-
-    let mut api_key = raw.api_key;
-    if let Some(val) = provider.api_key_env_var().and_then(env_non_empty) {
-        api_key = Some(val);
-    }
-
-    let base_url = raw
-        .base_url
-        .unwrap_or_else(|| provider.default_base_url().to_string());
-    let api_key = api_key.filter(|s| !s.is_empty());
-    let model = raw.model.filter(|s| !s.is_empty());
-
-    // Fall back to the provider's Dialogue preset if no model is configured.
-    // Skipped for Ollama: leaving model None lets setup_ollama_with_config pick
-    // a hardware-matched tier instead of the static preset tag.
-    let model = if provider.id() == "ollama" {
-        model
-    } else {
-        model.or_else(|| {
-            provider
-                .preset_model(InferenceCategory::Dialogue)
-                .map(String::from)
-        })
-    };
-
-    if provider.requires_api_key() && api_key.is_none() {
-        let hint = provider
-            .api_key_env_var()
-            .unwrap_or("the provider API key env var");
-        return Err(ParishError::Config(format!(
-            "{} provider requires an API key. Set {}.",
-            provider.id(),
-            hint
-        )));
-    }
-    if provider.needs_base_url_from_user() && base_url.is_empty() {
-        return Err(ParishError::Config(format!(
-            "{} provider requires a base_url. Set PARISH_BASE_URL or --base-url.",
-            provider.id()
-        )));
-    }
-
-    Ok(ProviderConfig {
-        provider,
-        base_url,
-        api_key,
-        model,
-    })
-}
-
-/// Resolves cloud provider configuration from file, env vars, and CLI flags.
-pub fn resolve_cloud_config(
-    config_path: Option<&Path>,
-    cli: &CliCloudOverrides,
-) -> Result<Option<CloudConfig>, ParishError> {
-    let toml_cfg = load_toml(config_path)?;
-
-    let toml_raw = RawLayers {
-        provider_str: toml_cfg.cloud.name,
-        base_url: toml_cfg.cloud.base_url,
-        api_key: toml_cfg.cloud.api_key,
-        model: toml_cfg.cloud.model,
-    };
-    let raw = apply_env_and_cli_layers(
-        toml_raw,
-        "PARISH_CLOUD",
-        cli.provider.as_deref(),
-        cli.base_url.as_deref(),
-        cli.model.as_deref(),
-    );
-
-    if raw.provider_str.is_none()
-        && raw.base_url.is_none()
-        && raw.api_key.is_none()
-        && raw.model.is_none()
-    {
-        return Ok(None);
-    }
-
-    // Default to OpenRouter for cloud. If the openrouter mod is absent
-    // from this deployment, surface that as a config error instead of
-    // panicking — operators who never set `PARISH_CLOUD_PROVIDER` should
-    // get an actionable message, not a crashed binary (codex P1).
-    let provider = match &raw.provider_str {
-        Some(s) => Provider::from_str_loose(s)?,
-        None => Provider::from_id("openrouter").ok_or_else(|| {
-            ParishError::Config(
-                "Cloud provider default 'openrouter' is not registered. \
-                 Set PARISH_CLOUD_PROVIDER (or [llm.cloud].provider) to a \
-                 registered provider id, or install the openrouter \
-                 provider mod under mods/openrouter-provider/."
-                    .into(),
-            )
-        })?,
-    };
-
-    let mut api_key = raw.api_key.filter(|s| !s.is_empty());
-    if let Some(val) = provider.api_key_env_var().and_then(env_non_empty) {
-        api_key = Some(val);
-    }
-
-    let base_url = raw
-        .base_url
-        .unwrap_or_else(|| provider.default_base_url().to_string());
-    let model = raw.model.filter(|s| !s.is_empty());
-
-    let model = model.ok_or_else(|| {
-        ParishError::Config(
-            "Cloud provider requires a model name. Set PARISH_CLOUD_MODEL or --cloud-model."
-                .to_string(),
-        )
-    })?;
-
-    if provider.requires_api_key() && api_key.is_none() {
-        let hint = provider
-            .api_key_env_var()
-            .unwrap_or("the provider API key env var");
-        return Err(ParishError::Config(format!(
-            "Cloud {} provider requires an API key. Set {}.",
-            provider.id(),
-            hint
-        )));
-    }
-    if provider.needs_base_url_from_user() && base_url.is_empty() {
-        return Err(ParishError::Config(format!(
-            "Cloud {} provider requires a base_url. Set PARISH_CLOUD_BASE_URL or --cloud-base-url.",
-            provider.id()
-        )));
-    }
-
-    Ok(Some(CloudConfig {
-        provider,
-        base_url,
-        api_key,
-        model,
-    }))
-}
-
-fn env_non_empty(key: &str) -> Option<String> {
-    std::env::var(key).ok().filter(|v| !v.is_empty())
-}
-
-fn read_toml_config(path: &Path) -> Result<TomlConfig, ParishError> {
-    if !path.exists() {
-        return Ok(TomlConfig::default());
-    }
-    let content = std::fs::read_to_string(path).map_err(|e| {
-        ParishError::Config(format!(
-            "failed to read config file {}: {}",
-            path.display(),
-            e
-        ))
-    })?;
-    toml::from_str(&content).map_err(|e| {
-        ParishError::Config(format!(
-            "failed to parse config file {}: {}",
-            path.display(),
-            e
-        ))
-    })
-}
+//!
+//! Structure (#1200 decomposition): the former single module is split into
+//! - [`category`] — the [`InferenceCategory`] enum;
+//! - [`schema`] — provider schema types (`ProviderKind`, `ProviderPreset`,
+//!   `ProviderMod`, …) and the [`Provider`] handle;
+//! - [`registry`] — the builtin/mod-loaded [`ProviderRegistry`] + accessors;
+//! - [`resolution`] — file/env/CLI config resolution.
+//!
+//! Every public item is re-exported flat here so the public paths
+//! (`provider::Provider`, `provider::resolve_config`, …, and the crate-root
+//! `pub use provider::*`) are unchanged.
+
+mod category;
+mod registry;
+mod resolution;
+mod schema;
+
+pub use category::InferenceCategory;
+pub use registry::{ProviderRegistry, ensure_mods_loaded, registry};
+pub use resolution::{
+    CliCloudOverrides, CliOverrides, CloudConfig, ProviderConfig, resolve_cloud_config,
+    resolve_config,
+};
+pub use schema::{
+    PresetBaseUrls, Provider, ProviderKind, ProviderMod, ProviderPreset, unified_memory_bytes,
+};
+
+// Private re-imports so the `tests` submodule (which uses `super::*`) keeps
+// reaching the crate-internal helpers it pins.
+#[cfg(test)]
+use resolution::read_toml_config;
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::builtin_providers;
     use serial_test::serial;
     use std::io::Write;
+    use std::path::Path;
 
     #[test]
     #[serial(provider_registry)]
